🚚 refactor Table component structure; migrate from foundations to molecules and enhance styling
