+++ packages/react/src/components/molecules/Table/styles/body.css.ts
