+++ packages/react/src/index.ts
@@ -28,12 +28,6 @@ export * from './components/foundations/Icon/index';
 export * from './components/foundations/Icon/shared';
 export * from './components/foundations/TAPIE-Symbol/index';
 export * from './components/foundations/TAPIE-Symbol/shared';
-export * from './components/foundations/Table/body/cell';
-export * from './components/foundations/Table/body/index';
-export * from './components/foundations/Table/body/row';
-export * from './components/foundations/Table/head/cell';
-export * from './components/foundations/Table/head/index';
-export * from './components/foundations/Table/index';
 export * from './components/foundations/Typography/builder';
 export * from './components/foundations/Typography/index';
 export * from './components/foundations/Typography/shared';
@@ -65,6 +59,12 @@ export * from './components/molecules/SegmentGroup/index';
 export * from './components/molecules/Sheet/header';
 export * from './components/molecules/Sheet/index';
 export * from './components/molecules/Sheet/shared';
+export * from './components/molecules/Table/body/cell';
+export * from './components/molecules/Table/body/index';
+export * from './components/molecules/Table/body/row';
+export * from './components/molecules/Table/head/cell';
+export * from './components/molecules/Table/head/index';
+export * from './components/molecules/Table/index';
 export * from './hooks/use-media-query';
 export * from './hooks/use-mounted-state';
 export * from './hooks/use-mouse-position';
